Rename nam.rs to chain.rs
