+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod aligner;
+pub mod chain;
 pub mod chainer;
 pub mod cigar;
 pub mod details;
@@ -10,7 +11,6 @@ pub mod maponly;
 pub mod mapper;
 pub mod math;
 pub mod mcsstrategy;
-pub mod nam;
 pub mod partition;
 pub mod piecewisealigner;
 pub mod read;
